Minor refactor on teams
